refactor: Move general stuff from test/utils.py to util/*
